requirements.txt
# local package
-e .

# external requirements
aiobotocore
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiohttp-retry==2.8.3
aioitertools==0.11.0
aiosignal==1.3.1
alembic==1.13.2
amqp==5.2.0
aniso8601==9.0.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appdirs==1.4.4
asyncssh==2.15.0
atpublic==5.0
attrs==23.2.0
awscli
backoff==2.2.1
billiard==4.2.0
blinker==1.8.2
boto3==1.34.131
botocore
cachetools==5.4.0
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.0.0
colorama==0.4.6
commonmark==0.9.1
configobj==5.0.8
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
dacite==1.6.0
dagshub==0.3.34
databricks-sdk==0.29.0
dataclasses-json==0.6.7
Deprecated==1.2.14
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docutils==0.16
dpath==2.2.0
dulwich==0.22.1
dvc==3.53.0
dvc-data==3.15.1
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-s3==3.2.0
dvc-studio-client==0.21.0
dvc-task==0.4.0
entrypoints==0.4
filelock==3.15.4
Flask==3.0.3
flatten-dict==0.4.2
flufl.lock==8.1.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
funcy==2.0
fusepy==3.0.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.32.0
gql==3.5.0
grandalf==0.8
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
gto==1.7.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
hydra-core==1.3.2
idna==3.7
importlib_metadata==7.2.1
iterative-telemetry==0.0.8
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.5
kombu==5.3.7
Mako==1.3.5
Markdown==3.6
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.7.5 
mlflow==2.15.0
mlflow-skinny==2.15.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
omegaconf==2.3.0
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pathvalidate==3.0.0
pillow==10.4.0
platformdirs==3.11.0
prompt_toolkit==3.0.47
protobuf==5.27.3
psutil==6.0.0
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydot==3.0.1
pygit2==1.15.1
Pygments==2.18.0
pygtrie==2.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
querystring-parser==1.2.4
regex==2024.7.24
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.1.0
rsa==4.7.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3fs==2024.6.1
s3transfer==0.10.2
scikit-learn==1.5.1
scipy==1.14.0
scmrepo==3.3.6
semver==3.0.2
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlparse==0.5.1
sqltrie==0.11.0
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.5.0
tomlkit==0.13.0
tqdm==4.66.4
treelib==1.6.4
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
voluptuous==0.15.2
waitress==3.0.0
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
zc.lockfile==3.0.post1
zipp==3.19.2
prometheus_client

-e .